增加删pycache的函数

+++ aiutils/fileutils.py
@@ -1,5 +1,5 @@
 #!/usr/bin/env python
-#coding:utf-8
+# coding:utf-8
 
 import json
 import os
@@ -9,7 +9,25 @@ import math
 
 
 def remove_folder(path):
-    shutil.rmtree(path)
+    if isExist(path):
+        shutil.rmtree(path)
+
+
+def remove_pycaches(dir):
+    need_deletes = []
+    for home, dirs, files in os.walk(dir):
+        for dir in dirs:
+            fullname = os.path.join(home, dir)
+            if dir == "__pycache__":
+                need_deletes.append(fullname)
+            else:
+                pass
+
+        for filename in files:
+            fullname = os.path.join(home, filename)
+
+    for need_del in need_deletes:
+        remove_folder(need_del)
 
 
 def remove_pre_checkpoints(model_path):
@@ -21,26 +39,31 @@ def remove_pre_checkpoints(model_path):
         if tfile.find("step_") >= 0:
             steps.append([tfile, int(tfile.replace("step_", ""))])
     steps.sort(key=lambda x: x[1])
-    if len(steps)<1:
+    if len(steps) < 1:
         return
     steps.pop()
     for step in steps:
         print("del", step[0])
         remove_folder(model_path + "/" + step[0])
 
+
 def ensure_dir(dir_path):
     if isExist(dir_path):
         return
     os.makedirs(dir_path)
 
-def copy(file_path,new_path):
+
+def copy(file_path, new_path):
     shutil.copy(file_path, new_path)
+
 
 def isExist(fPath):
     return os.path.exists(fPath)
 
+
 def file_extension(path):
-  return os.path.splitext(path)[1].replace(".","")
+    return os.path.splitext(path)[1].replace(".", "")
+
 
 def removeFile(fPath):
     if not isExist(fPath):
@@ -63,60 +86,69 @@ def readJsonFile(filepath):
     # print(fc)
     return json.loads(fc)
 
+
 def readFile(filepath):
-    f=open(filepath,"r",encoding='utf-8')
-    fc=f.read()
+    f = open(filepath, "r", encoding='utf-8')
+    fc = f.read()
     f.close()
-    #print(fc)
-    return fc        
-def saveFile(filepath,content):
-    f=open(filepath,"w",encoding='utf-8')
-    fc=f.write(content)
-    f.close()
-    #print(fc)
+    # print(fc)
     return fc
 
+
+def saveFile(filepath, content):
+    f = open(filepath, "w", encoding='utf-8')
+    fc = f.write(content)
+    f.close()
+    # print(fc)
+    return fc
+
+
 def adptSHFile(filePath):
-    txt=readFile(filePath)
-    txt=txt.replace("\r","")
-    saveFile(filePath,txt)
+    txt = readFile(filePath)
+    txt = txt.replace("\r", "")
+    saveFile(filePath, txt)
+
 
 def adptSHFiles(fileList):
     for tfile in fileList:
-        adptSHFile(tfile)   
+        adptSHFile(tfile)
+
 
 def saveJsonLines(path, data):
-    lines=[]
+    lines = []
     for line in data:
         lines.append(json.dumps(line, ensure_ascii=False))
-        
-    content="\n".join(lines)
-    saveFile(path,content)
 
-def saveJsonLines2(path,data):
+    content = "\n".join(lines)
+    saveFile(path, content)
+
+
+def saveJsonLines2(path, data):
     out_file = open(
         path, 'w', encoding='utf-8'
     )
     for line in data:
-        out_file.write(json.dumps(line,ensure_ascii=False)+"\n")
+        out_file.write(json.dumps(line, ensure_ascii=False) + "\n")
     out_file.close()
 
+
 def readJsonLines(filepath):
-    lines=readFile(filepath).split("\n")
-    linefiles=[]
+    lines = readFile(filepath).split("\n")
+    linefiles = []
     for line in lines:
 
-        line=line.strip()
+        line = line.strip()
         if not line:
             continue
-        dd=json.loads(line)
+        dd = json.loads(line)
         linefiles.append(dd)
     return linefiles
 
+
 def readLines(dataFile):
-    txt=readFile(dataFile)
-    lines=txt.split("\n")
-    rst=[]
+    txt = readFile(dataFile)
+    lines = txt.split("\n")
+    rst = []
     for line in lines:
         if not line:
             continue
@@ -124,20 +156,19 @@ def readLines(dataFile):
     return rst
 
 
-def sampleFileLines(dataFile,savePath,rate=0.1,hasHead=False,useRandom=True,count=-1):
-    lines=readLines(dataFile)
+def sampleFileLines(dataFile, savePath, rate=0.1, hasHead=False, useRandom=True, count=-1):
+    lines = readLines(dataFile)
     if hasHead:
-        title=lines[0]
-        lines=lines[1:]
+        title = lines[0]
+        lines = lines[1:]
 
     if useRandom:
         random.shuffle(lines)
 
-    if count<1:
+    if count < 1:
         count = math.floor(len(lines) * rate)
 
-    sampled= lines[0:count]
+    sampled = lines[0:count]
     if hasHead:
-        sampled=[title]+sampled
-    saveFile(savePath,"\n".join(sampled))
-
+        sampled = [title] + sampled
+    saveFile(savePath, "\n".join(sampled))
